+++ files/settings.py
@@ -1,0 +1,8 @@
+# File settings
+ino_file = 1
+file_extension = ".txt"  # Only when ino_file = 0
+
+# Auto delay settings (Under review)
+# auto_delay_toggle = 0
+# enter_delay = "0"
+# win_r_delay = "0"