+++ settings.py
@@ -1,3 +1,0 @@
-ENTER_DELAY = "0"
-WIN_R_DELAY = "0"
-file_extension = ".txt"